backend/requirements.txt
fastapi
uvicorn
python-socketio[asyncio_client]
python-socketio[asyncio_server]